Playing with the implementation

+++ src/serial/mod.rs
@@ -1,9 +1,9 @@
 pub mod uart;
 pub mod usart;
 
-pub struct BaudRate(pub u16);
+pub struct BaudRate(pub u32);
 
-impl Into<BaudRate> for u16 {
+impl Into<BaudRate> for u32 {
     fn into(self) -> BaudRate {
         BaudRate(self)
     }
